entrez-scripts/requirements.txt
biopython==1.64
progressbar33==2.4